backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "alpharag"
version = "0.1.0"
description = "Financial research agent over SEC filings with citation-backed answers"
readme = "../README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "AlphaRAG" }]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "alembic>=1.14",
    "pgvector>=0.3.6",
    "httpx>=0.27",
    "selectolax>=0.3.25",
    "tiktoken>=0.8",
    "openai>=1.54",
    "structlog>=24.4",
    "sse-starlette>=2.1",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-PyYAML",
    "PyYAML>=6.0",
]

[project.scripts]
alpharag-warmup = "scripts.warmup:main"
alpharag-search = "scripts.search:main"

[tool.hatch.build.targets.wheel]
packages = ["src/alpharag"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM", "RUF"]
ignore = ["E501"]  # handled by formatter

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "N"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "integration: tests that require a running Postgres",
    "eval: end-to-end eval suite (slow, costs money)",
]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["tests/", "alembic/"]
